Fix sortpom consistency (#39)
Ensure consistency of space before an empty element is closed. This
fixes conflicts between different tools, some of which add the space
(like maven-release-plugin) and the sortpom plugin, which previously
removed it, by making the sortpom plugin always add it.

This ensures the `<mailingList />` doesn't keep changing to
`<mailingList/>` back and forth.